requirements.txt
PySide6==6.8.0.2
requests==2.31.0
python-dotenv==1.0.1
python-docx==1.1.0
ebooklib==0.18
beautifulsoup4==4.12.3
